refactor: move and clean counsellor dasboard components (#20)
