+++ frontend/src/lib/components/dashboard/ModuleCircleButton.svelte
