+++ frontend/src/lib/components/dashboard/ModuleProgress.svelte
@@ -1,5 +1,8 @@
 <script lang="ts">
-  let modPrint = 'Completed : 10 / 40';
+  export let completed = 10;
+  export let total = 40;
+
+  let percent = completed / total * 100;
 </script>
 
 <!-- <progress class="progress  w-56 h-5 align-middle bg-base-400" value="70" max="100"></progress> -->
@@ -8,10 +11,10 @@
   <div class="w-full bg-gray-200 rounded-full">
     <div
       class="bg-secondary text-xs font-medium text-blue-100 text-center p-0.5 leading-none rounded"
-      style="width: 25%"
+      style="width: {percent}%"
     >
-      25%
+      {percent}%
     </div>
   </div>
-  <div class="font-xs text-base-400">{modPrint}</div>
+  <div class="font-xs text-base-400">Completed: {completed}/{total}</div>
 </div>
