--- frontend/src/lib/components/moduleProgress.svelte
@@ -1,12 +1,0 @@
-<script lang="ts">
-</script>
-
-<!-- <progress class="progress  w-56 h-5 align-middle bg-base-400" value="70" max="100"></progress> -->
-<div class="w-96 h-6 bg-base-300 rounded-md m-3">
-  <div
-    class="h-6 bg-blue-light text-sm font-medium text-white text-center p-1 leading-none rounded-md"
-    style="width: 45%"
-  >
-    45%
-  </div>
-</div>
